Move more of the settings infrastructure to ILSpyX for reuse.
