+++ ICSharpCode.ILSpyX/Settings/DecompilerSettings.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) 2024 Tom Englert for the SharpDevelop Team
+﻿// Copyright (c) 2024 Tom Englert
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this
 // software and associated documentation files (the "Software"), to deal in the Software
@@ -21,9 +21,7 @@ using System.Linq;
 using System.Reflection;
 using System.Xml.Linq;
 
-#nullable enable
-
-namespace ICSharpCode.ILSpy.Options
+namespace ICSharpCode.ILSpyX.Settings
 {
 	public class DecompilerSettings : Decompiler.DecompilerSettings, ISettingsSection
 	{
