+++ ILSpy/Options/DisplaySettings.cs
@@ -19,6 +19,8 @@
 using System.Windows.Media;
 using System.Xml.Linq;
 
+using ICSharpCode.ILSpyX.Settings;
+
 using TomsToolbox.Wpf;
 
 namespace ICSharpCode.ILSpy.Options
